backend/requirements.txt
Django
gunicorn
whitenoise
numpy
pillow
tensorflow
dj-database-url
psycopg2-binary
python-dotenv